# Conflicts:
#	Service.API/DataAccess/DataAccess/Models/MSSQL/Fantasy/Model.edmx
#	Service.API/DataAccess/DataAccess/Models/MSSQL/Fantasy/Model.edmx.sql
#	Service.API/DataAccess/DataAccess/Services/Fantasy/Core/DatabaseCore.cs

+++ Service.API/Domain/Domain/BussinessObjects/FantasyBOs/ContestBO.cs
@@ -8,7 +8,7 @@ namespace Fantasy.API.Domain.BussinessObjects.FantasyBOs
 {
     public class ContestBO
     {
-        public int ContestId { get; set; }
+        public long ContestId { get; set; }
 
         public ContestTypeBO ContestTypeId { get; set; }
